fix: skipping default merge commit messages
Currently, merge commits are validated for whether they contain a Rally story,
defect, or task. For scenarios where the branch name contains a Rally ticket
identifier, when that branch is merged, the branch name containing the Rally
ticket identifier will show up in the merge commit message and fail
validation as it may be missing a #.  This changeset will omit merge commits
with the default merge commit message from being validated for Rally ticket
identifiers.